pyproject.toml
[tool.poetry]
name = "catservable"
version = "0.1.0"
description = ""
authors = ["decko <brito.afa@gmail.com>"]
license = "GPl-2.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.5"
djangorestframework = "^3.12.2"
django-prometheus = "^2.1.0"
psycopg2 = "^2.8.6"
prettyconf = "^2.2.1"
dj-database-url = "^0.5.0"
model-bakery = "^1.2.1"
tapioca-wrapper = "^2.0.1"
jedi = "0.17.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pytest-django = "^4.1.0"
ipdb = "^0.13.4"
ipython = "^7.19.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "catservable.settings"
addopts = "--reuse-db"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
